Add after callback to be called at the end of the addHandler call
This commit adds an `after` option to the object passed as the second
argument to  `enquire.register`. If supplied, it will be triggered
once immediately after registration of the handler.

This allows you to execute any code after the handler has been added,
regardless of whether the handler was called or not.

In my use case, I'm using enquire to move a block of content into a
sidebar if a media query matches. I needed to execute some code after
the block was either moved or wasn't, and this allows me to do that.